Adding 3 smashed commits into this one. Deleted `highlight_model_100e.h5` as it overpasses the 100mb size limit. Also updated most preproc methods to match current configuration and finished the demo (YAY!!). Added some models. Also improved line segmentation inference speed in 6x!! Now everything finally works.

+++ preproc/RobustBinarize.py
@@ -76,7 +76,7 @@ class RobustBinarize:
 
         # 5. Compute result
         result = RobustBinarize.compute_result(cei, erosion, ldi, bl)
-        return result
+        return result.astype(np.uint8)
 
     @staticmethod
     @nb.njit(cache=True)
